+++ local_pxs_vs_rsync.sh
@@ -80,7 +80,7 @@ main() {
     
     setup_baseline
     overwrite_middle
-    run_bench "Aligned-Overwrite" "pxs" "$PXS_CMD sync $DST $SRC ${PXS_FLAGS[*]}"
+    run_bench "Aligned-Overwrite" "pxs" "$PXS_CMD sync $SRC $DST ${PXS_FLAGS[*]}"
 
     # Scenario B: Data Shift
     setup_baseline
@@ -89,7 +89,7 @@ main() {
     
     setup_baseline
     prepend_byte
-    run_bench "Data-Shift" "pxs" "$PXS_CMD sync $DST $SRC ${PXS_FLAGS[*]}"
+    run_bench "Data-Shift" "pxs" "$PXS_CMD sync $SRC $DST ${PXS_FLAGS[*]}"
 
     echo -e "\nBenchmark complete."
 }